issue #117 refactor: group member component separate
